+++ .travis.yml
@@ -1,7 +1,6 @@
 language: python
 
 python:
-  - "3.4"
   - "3.5"
   - "3.6"
 
@@ -13,7 +12,7 @@ script:
   - flake8 --ignore=E501, W605
   - autopep8 -r . --diff | tee check_autopep8
   - test ! -s check_autopep8
-  - python atc_env.py arc050 --without-login
+  - python envgen.py arc050 --without-login
 
 notifications:
   email: false
