+++ atcodertools/tools/envgen.py
